refactor: DomainTest 세분화
1. CarTest - 자동차 전진 및 현재 상태 테스트
2. RacingCarGameTest - 가장 멀리간 자동차 및 우승한 자동차 테스트
