+++ src/test/kotlin/RacingCarGameTest.kt
@@ -1,21 +1,10 @@
 import model.Car
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.CsvSource
 import service.RacingCarGameService
 
-class DomainTest {
+class RacingCarGameTest {
     private val racingCarGameService = RacingCarGameService()
-
-    @DisplayName("차의 현재 상태를 잘 출력하는지 확인하는 테스트")
-    @ParameterizedTest
-    @CsvSource("aa,3,aa : ---", "bb,2,bb : --")
-    fun testCarPresentStatus(name: String, position: Int, expected: String) {
-        val actual = Car(name, position).getPresentStatus()
-        assertThat(actual).isEqualTo(expected)
-    }
 
     @Test
     fun `가장 멀리간 자동차를 확인하는 테스트`() {
